requirements.txt
requests==2.27.1
python-telegram-bot==13.11
